==> jupyter_notebooks/04 - DataVisualization.ipynb <==
# Data Visualization Notebook

## Objectives

*   Answer business requirement 3: 
    * As a customer I am interested to understand the rainfall seasonality for a given city in the last 5 years.


## Inputs

* outputs/datasets/collection/WeatherAustralia.csv

## Outputs

* generate code that answers business requirement 3 and can be used to build Streamlit App

## Additional Comments | Insights | Conclusions




---

# Install Packages

! pip install matplotlib -U
! pip install pandas-profiling==2.11.0
! pip install plotly==4.14.0

# Code for restarting the runtime, that will restart colab session
# It is a good practice after you install a package in a colab session
import os
os.kill(os.getpid(), 9)

# Setup GPU

* Go to Edit → Notebook Settings
* In the Hardware accelerator menu, selects GPU
* note: when you select an option, either GPU, TPU or None, you switch among kernels/sessions

---
* How to know if I am using the GPU?
  * run the code below, if the output is different than '0' or null/nothing, you are using GPU in this session
  * Typically the output will be /device:GPU:0


import tensorflow as tf
tf.test.gpu_device_name()

# **Connection between: Colab Session and your GitHub Repo**

### Insert your **credentials**

* The variable's content will exist only while the session exists. Once this session terminates, the variable's content will be erased permanently.

from getpass import getpass
import os
from IPython.display import clear_output 

print("=== Insert your credentials === \nType in and hit Enter")
os.environ['UserName'] = getpass('GitHub User Name: ')
os.environ['UserEmail'] = getpass('GitHub User E-mail: ')
os.environ['RepoName'] = getpass('GitHub Repository Name: ')
os.environ['UserPwd'] = getpass('GitHub Account Password: ')
clear_output()
print("* Thanks for inserting your credentials!")
print(f"* You may now Clone your Repo to this Session, "
      f"then Connect this Session to your Repo.")

---

### **Clone** your GitHub Repo to your current Colab session

* So you can have access to your project's files

! git clone https://github.com/{os.environ['UserName']}/{os.environ['RepoName']}.git
! rm -rf sample_data   # remove content/sample_data folder, since we dont need it for this project

import os
if os.path.isdir(os.environ['RepoName']):
  print("\n")
  %cd /content/{os.environ['RepoName']}
  print(f"\n\n* Current session directory is:{os.getcwd()}")
  print(f"* You may refresh the session folder to access {os.environ['RepoName']} folder.")
else:
  print(f"\n* The Repo {os.environ['UserName']}/{os.environ['RepoName']} was not cloned."
        f" Please check your Credentials: UserName and RepoName")

---

### **Connect** this Colab session to your GitHub Repo

* So if you need, you can push files generated in this session to your Repo.

! git config --global user.email {os.environ['UserEmail']}
! git config --global user.name {os.environ['UserName']}
! git remote rm origin
! git remote add origin https://{os.environ['UserName']}:{os.environ['UserPwd']}@github.com/{os.environ['UserName']}/{os.environ['RepoName']}.git

# the logic is: create a temporary file in the sessions, update the repo. Delete this file, update the repo
# If it works, it is a signed that the session is connected to the repo.
import uuid
file_name = "session_connection_test_" + str(uuid.uuid4()) # generates a unique file name
with open(f"{file_name}.txt", "w") as file: file.write("text")
print("=== Testing Session Connectivity to the Repo === \n")
! git add . ; ! git commit -m {file_name + "_added_file"} ; ! git push origin main 
print("\n\n")
os.remove(f"{file_name}.txt")
! git add . ; ! git commit -m {file_name + "_removed_file"}; ! git push origin main

# delete your Credentials (username and password)
os.environ['UserName'] = os.environ['UserPwd'] = os.environ['UserEmail'] = ""

* If output above indicates there was a **failure in the authentication**, please insert again your credentials.

---

### **Push** generated/new files from this Session to GitHub repo

* Git status

! git status

* Git commit

CommitMsg = "added-cleaned-data"
!git add .
!git commit -m {CommitMsg}

* Git Push

!git push origin main


---

### **Delete** Cloned Repo from current Session

%cd /content
!rm -rf {os.environ['RepoName']}

print(f"\n * Please refresh session folder to validate that {os.environ['RepoName']} folder was removed from this session.")
print(f"\n\n* Current session directory is:  {os.getcwd()}")

---

# Load your data

import pandas as pd
df = pd.read_csv("/content/WalkthroughProject/outputs/datasets/collection/WeatherAustralia.csv")

df['Date'] = pd.to_datetime(df['Date'])
df['Day'] = df['Date'].dt.day
df['Month'] = df['Date'].dt.month
df['Year'] = df['Date'].dt.year
df['WeekDay']=df['Date'].dt.weekday
df['IsWeekend'] = df['WeekDay'].apply(lambda x: 1 if x >= 5 else 0)
df['YearMonth'] = df['Date'].dt.to_period('M')

df['WeekDay']=df['Date'].dt.day_name() # gets day name 
days_order = ["Monday","Tuesday","Wednesday","Thursday","Friday","Saturday","Sunday"]
df['WeekDay'] = pd.Series(df['WeekDay'], dtype=pd.CategoricalDtype(categories=days_order, ordered=True))

df.set_index(['Date'],drop=True,inplace=True)
df.head(3)

# Quick exploration with Pandas Profiling

from pandas_profiling import ProfileReport
pandas_report = ProfileReport(df=df,minimal=True)
pandas_report.to_notebook_iframe()

# Data Visualization Tasks

* Subset relevant variables for analysis

df_vis = df.filter(['Location','RainfallToday','Month', 'Year','YearMonth', 'WeekDay','State']).copy()
df_vis.head(3)

## Business Requirement 3

* As a customer I am interested to understand the rainfall seasonality for a given city in the last 5 years.

* We will subset a given city and the data from the last 5 years

df_vis['Location'].unique()

city = 'Canberra'
years_backward = 5

df_city = df_vis.query(f"Location == '{city}'")
df_city = df_city.query(f"Year > {df_city['Year'].max() - years_backward}").drop(['Location'],axis=1)

print(f"* City: {city} \n* Index min: {df_city.index.min()} \n* Index max: {df_city.index.max()}")

* Week day Seasonality

import plotly.express as px

df_day = df_city.filter(['RainfallToday','WeekDay']).groupby(by=['WeekDay']).agg('mean')
fig = px.line(df_day, x=df_day.index, y='RainfallToday',title=f'Week Day Seasonality in {city}')
fig.update_xaxes(type='category')
fig.update_yaxes(title='Raifall Levels',showticklabels=False)  # show or not??
fig.show()

* Monthly seasonality

df_month = df_city.filter(['RainfallToday','Month']).groupby(by=['Month']).agg('mean')
fig = px.line(df_month, x=df_month.index, y='RainfallToday',title=f'Rainfall - Monthly Seasonality in {city}')
fig.update_xaxes(type='category')
fig.update_yaxes(title='Raifall Levels',showticklabels=False)
fig.show()

* Yearly Seasonality

df_year = df_city.filter(['RainfallToday','Year']).groupby(by=['Year']).agg('mean')
fig = px.line(df_year, x=df_year.index, y='RainfallToday',title=f'Rainfall - Yearly Seasonality in {city}')
fig.update_xaxes(type='category')
fig.update_yaxes(title='Raifall Levels',showticklabels=False)
fig.show()

* Avg Rainfall Levels per state

df_state = df_vis.filter(['RainfallToday','State']).groupby(by=['State']).agg('mean')
fig = px.bar(df_state, x=df_state.index, color=df_state.index,y='RainfallToday',
             title=f'Rainfall - State Seasonality')
fig.update_xaxes(type='category')
fig.update_yaxes(title='Raifall Avg Levels')
fig.show()

## Additional Analysis

* Subset neeeded variables


variables_eda = ['RainfallToday','RainToday',
                 'Latitude','Longitude','Location', 'State',
                 'Day', 'Month', 'Year','YearMonth','WeekDay', 'IsWeekend']
df_eda = df.filter(variables_eda).copy()

years_backward = 5
df_eda = df_eda.query(f"Year > {df_eda['Year'].max() - years_backward}")

df_eda.head(3)

* Plots we are interested
   * map
   * heatmap
   * pairplot

### Map

* for a given year, animate by month, agg mean levels of RainfallToday

map_year = [2017]

df_map= df_eda.query(f"Year in {map_year}").copy()
df_map_month = df_map[['RainfallToday','Location','Month','Year']].groupby(['Location','Month','Year']).mean().reset_index()


df_map_month=(df_map_month
              .merge(df_map[['Location','State','Latitude','Longitude']],how='right',on='Location')
              .sort_values(by=['Location','Month','Year'])
              .drop_duplicates()
              )
df_map_month

import plotly.express as px 
fig = px.scatter_mapbox(df_map_month.dropna(),
                        lat="Latitude", lon="Longitude", color="State",
                        # hover_data=["RainfallToday",'Location'],
                        size='RainfallToday',
                        zoom=2.5,
                        mapbox_style="open-street-map",
                        animation_frame='Month',
                        animation_group = 'Year',
                        hover_data = ['Location', 'Month', 'Year'],
                        center={"lat":-27,"lon":133},
                        size_max=25,width=None, height=600
                        )
fig.show()

### Rainfall Heatmap

* barplot indicating rainToday flag, for given city, over time

! pip install calplot==0.1.7.2

df_eda['Location'].unique()

city = 'Sydney'
df_rain_flag= df_eda.query(f"Location == '{city}' ").copy()
df_rain_flag.head()

import matplotlib.pyplot as plt
import calplot
for year_heatmap in df_rain_flag['Year'].unique():
  print(f"\n * {year_heatmap} \n")
  fig= plt.figure(figsize=(20,5))
  calplot.yearplot(data=df_rain_flag.query(f"Year == {year_heatmap} ")['RainfallToday'],
                  dropzero= True,
                  cmap='GnBu',
                  linewidth =2,
                  # fillcolor='black'
                  );
  plt.show()

### Boxplot

df_eda.head()

df_rain_flag['Year'].unique()

city = 'Sydney'
df_rain_flag= df_eda.query(f"Location == '{city}' ").copy()
df_rain_flag.head()

select_x='Month'

fig = px.box(df_rain_flag, x=select_x, y='RainfallToday',color=select_x) #,log_y=True)
fig.update_layout(xaxis_type = 'category')
fig.show()

### Other plots

# histograms, per categorical
 # rainfall x jan/2013 vs jan2014 etc
  # rainfall x years
  

# pairpolot scatter plot
  # RainfallToday		Latitude	Longitude
  # color by state, year,
